structural changes of the folders and files
