+++ src/main.tsx
@@ -6,7 +6,7 @@ import { Provider } from "react-redux";
 import store from "./store/index.tsx";
 import i18next from "i18next";
 import signIn from "./utilities/en/signIn.json";
-import appointmentCard from "./utilities/en/AppointmentCard.json";
+import status from "./utilities/en/Status.json";
 import { I18nextProvider } from "react-i18next";
 
 // <> Language translation feature
@@ -16,7 +16,7 @@ i18next.init({
   resources: {
     en: {
       signIn: signIn,
-      appointmentCard: appointmentCard,
+      status: status,
     },
   },
 });
