--- src/utilities/en/AppointmentCard.json
@@ -1,8 +1,0 @@
-{
-  "appointmentCard": {
-    "completed": "Completed",
-    "inProgress": "In-Progress",
-    "scheduled": "Scheduled",
-    "cancelled": "Cancelled"
-  }
-}
